Reorganize and prepare for next version
