+++ src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
 name: skript-placeholders
-main: com.apickledwalrus.papiaddon.Main
-version: 1.2
+main: io.github.apickledwalrus.placeholderaddon.Main
+version: 1.3
 authors: [Pikachu, APickledWalrus]
 description: Allows using and registering placeholders with PlaceholderAPI in Skript
 website: https://github.com/APickledWalrus/skript-placeholders
